code cleanup, collect remaining info confirm room
Change so the list adapters don't require a layout and text view, since
they're only made to work with a single layout anyway we can just define
it as a constant within the class.

ConfirmRoomFragment has been reorganized so all information from the
form is captured when clicking ReserveRoom, and this is done in the
order the information appears in the form.