add booking coordinator and start migrating
The booking coordinator replaces the use of the booking bundle which we
thought, early on in the project, was necessary for passing data between
the fragments. However since the fragments share a single activity root
through which they can exchange unparceled data, this solution is much
cleaner and better.

As of yet only the date fragment has fully migrated, though search
result fragment is almost there as well.